backend/requirements/common.txt
Django>=4.2,<5

# for image handling
pillow==10.3.0

# for audio compression
pydub==0.25.1

# for REST API support
djangorestframework==3.15.1

# for JWT based authentication
djangorestframework-simplejwt==5.3.1

# PostgreSQL database adapter for Django
psycopg[binary]==3.1.18


# for s3 bucket storage
django-storages[s3]==1.14.3

# python aws sdk
boto3==1.34.106
boto3-stubs[essential]==1.34.142


# for background tasks
celery[redis]==5.4.0

